Adds redis broker
This commit adds a redis broker to the mix. The setup now has two
logstash components: a log shipper -  which parses and sends the logs on
to Redis, and the indexer - which recieves the logs from Redis and sends
them on to ElasticSearch.

The Redis broker container has an exposed port so that (in theory) it
can receive logs shipped from other hosts.